+++ Cargo.toml
@@ -31,6 +31,7 @@ members = [
     "kernel/libs/typeflags",
     "kernel/libs/typeflags-util",
     "kernel/libs/atomic-integer-wrapper",
+    "kernel/libs/aster-page-allocator",
 ]
 exclude = [
     "kernel/libs/comp-sys/cargo-component",
