+++ osdk/src/base_crate/main.rs.template
@@ -1,7 +1,12 @@
 #![no_std]
 #![no_main]
+#![feature(linkage)]
 
 extern crate #TARGET_NAME#;
+extern crate alloc;
+
+use ostd::mm::page::allocator::PageAlloc;
+use alloc::boxed::Box;
 
 #[panic_handler]
 fn panic(info: &core::panic::PanicInfo) -> ! {
@@ -9,4 +14,9 @@ fn panic(info: &core::panic::PanicInfo) -> ! {
         pub fn __ostd_panic_handler(info: &core::panic::PanicInfo) -> !;
     }
     unsafe { __ostd_panic_handler(info); }
+}
+
+#[ostd::default_page_allocator_init_fn]
+fn init_page_allocator() -> Box<dyn PageAlloc> {
+    aster_page_allocator::init()
 }
