+++ osdk/test-kernel/src/lib.rs
@@ -6,6 +6,7 @@
 
 #![no_std]
 #![forbid(unsafe_code)]
+#![feature(linkage)]
 
 extern crate alloc;
 
@@ -20,6 +21,7 @@ use ostd::{
     ktest::{
         get_ktest_crate_whitelist, get_ktest_test_whitelist, KtestError, KtestItem, KtestIter,
     },
+    mm::page::allocator::PageAlloc,
 };
 use owo_colors::OwoColorize;
 use path::{KtestPath, SuffixTrie};
@@ -74,6 +76,11 @@ fn panic_handler(info: &core::panic::PanicInfo) -> ! {
     early_println!("An uncaught panic occurred: {:#?}", throw_info);
 
     ostd::prelude::abort();
+}
+
+#[ostd::ktest::page_allocator_init_fn]
+fn init_page_allocator() -> Box<dyn PageAlloc> {
+    aster_page_allocator::init()
 }
 
 /// Run all the tests registered by `#[ktest]` in the `.ktest_array` section.
